+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on the [KeepAChangeLog] project.
 - [#430] Audience of a client assertion is endpoint dependent.
 - [#427] Made matching for response_types order independent for authorization requests
 - [#399] Matching response_types for authz requests is too strict
+- [#xxx] Fixed client.read_registration
 
 [#430]: https://github.com/OpenIDC/pyoidc/pull/430
 [#427]: https://github.com/OpenIDC/pyoidc/pull/427